+++ wp-content/plugins/redis-cache-pro/src/Metrics/WordPressMetrics.php
@@ -105,11 +105,11 @@ class WordPressMetrics
     public $msCache;
 
     /**
-     * The median amount of time (ms) waited for the datastore to respond.
+     * The average amount of time (ms) waited for the datastore to respond.
      *
      * @var float
      */
-    public $msCacheMedian;
+    public $msCacheAvg;
 
     /**
      * The percentage of time waited for the datastore to respond,
@@ -140,8 +140,8 @@ class WordPressMetrics
         $this->storeWrites = $metrics->storeWrites;
         $this->storeHits = $metrics->storeHits;
         $this->storeMisses = $metrics->storeMisses;
-        $this->msCache = round($metrics->storeWait * 1000, 2);
-        $this->msCacheMedian = round($metrics->storeWaitAverage * 1000, 2);
+        $this->msCache = round($metrics->storeWait, 2);
+        $this->msCacheAvg = round($metrics->storeWaitAverage, 4);
 
         $requestStart = $_SERVER['REQUEST_TIME_FLOAT'] ?? $timestart;
 
@@ -175,7 +175,7 @@ class WordPressMetrics
             'sql-queries' => $this->sqlQueries,
             'ms-total' => sprintf('%.2f', $this->msTotal),
             'ms-cache' => sprintf('%.2f', $this->msCache),
-            'ms-cache-median' => sprintf('%.2f', $this->msCacheMedian),
+            'ms-cache-avg' => sprintf('%.4f', $this->msCacheAvg),
             'ms-cache-ratio' => number_format($this->msCacheRatio, 1),
         ];
     }
@@ -262,9 +262,9 @@ class WordPressMetrics
                 'description' => 'The total amount of time (ms) waited for the datastore to respond.',
                 'type' => 'time',
             ],
-            'ms-cache-median' => [
+            'ms-cache-avg' => [
                 'title' => 'Datastore Command Time',
-                'description' => 'The median amount of time (ms) waited for the datastore to respond.',
+                'description' => 'The average amount of time (ms) waited for the datastore to respond.',
                 'type' => 'time',
             ],
             'ms-cache-ratio' => [