+++ wp-content/plugins/redis-cache-pro/src/Plugin/Analytics.php
@@ -153,7 +153,7 @@ trait Analytics
         printf(
             "\n<!-- plugin=%s client=%s %s -->\n",
             'object-cache-pro',
-            $wp_object_cache->clientName(),
+            strtolower($wp_object_cache->clientName()),
             (string) $measurement
         );
     }