+++ wp-content/plugins/redis-cache-pro/src/Plugin/Extensions/QueryMonitor.php
@@ -38,6 +38,8 @@ trait QueryMonitor
         }
 
         add_action('init', [$this, 'registerQmCollectors']);
+        add_action('qm/constants', [$this, 'registerQmConstants']);
+
         add_filter('qm/outputter/html', [$this, 'registerQmOutputters']);
 
         add_filter('qm/component_type/unknown', [$this, 'fixUnknownQmComponentType'], 10, 2);
@@ -47,6 +49,27 @@ trait QueryMonitor
 
         add_filter('qm/component_context/plugin', [$this, 'fixUnknownQmComponentContext'], 10, 2);
         add_filter('qm/component_context/mu-plugin', [$this, 'fixUnknownQmComponentContext'], 10, 2);
+    }
+
+    /**
+     * Registers all object cache related Query Monitor constants.
+     *
+     * @param  array<string, array<string, mixed>>  $constants
+     * @return array<string, array<string, mixed>>
+     */
+    public function registerQmConstants($constants)
+    {
+        $constants['QM_OBJECTCACHE_EXPENSIVE'] = [
+            'label' => 'If an individual cache command takes longer than this time to execute, it’s considered "slow" and triggers a warning.',
+            'default' => 0.005,
+        ];
+
+        $constants['QM_OBJECTCACHE_HEAVY'] = [
+            'label' => 'If an individual cache key is larger than this in bytes, it’s considered "heavy" and triggers a warning.',
+            'default' => 1024 * 1024,
+        ];
+
+        return $constants;
     }
 
     /**
@@ -80,7 +103,7 @@ trait QueryMonitor
         }
 
         // Added in Query Monitor 3.1.0
-        if (! method_exists('QM_Output_Html', 'before_non_tabular_output')) {
+        if (! method_exists('QM_Output_Html', 'before_non_tabular_output')) { // @phpstan-ignore-line
             return;
         }